fix(anyharness): align terminal split after live boundary merge
