+++ anyharness/crates/anyharness-lib/src/workspaces/retention.rs
@@ -7,7 +7,7 @@ use std::time::Duration;
 use anyharness_contract::v1::{WorkspaceRetireBlocker, WorktreeRetentionRowOutcome};
 
 use crate::sessions::store::SessionStore;
-use crate::terminals::store::TerminalStore;
+use crate::domains::terminals::store::TerminalStore;
 use crate::workspaces::checkout_gate::{CheckoutDeletionGate, CheckoutPathLockKey};
 use crate::workspaces::managed_root::canonical_managed_worktrees_root;
 use crate::workspaces::model::WorkspaceRecord;
