+++ scripts/anyharness_boundaries_allowlist.txt
@@ -6,9 +6,6 @@
 # This is migration debt, not permission for new code. If a cleanup reduces a
 # count, update this file in the same PR.
 
-API_LIVE_RUNTIME_IMPORT anyharness/crates/anyharness-lib/src/api/http/terminals.rs 9 transitional terminal model/service types still live under top-level terminals
-API_LIVE_RUNTIME_IMPORT anyharness/crates/anyharness-lib/src/api/http/workspaces_contract.rs 2 transitional workspace contract mapper still reads terminal model types from top-level terminals
-API_LIVE_RUNTIME_IMPORT anyharness/crates/anyharness-lib/src/api/ws/terminals.rs 2 transitional terminal websocket still reads terminal runtime helpers from top-level terminals
 SESSION_STORE_LIVE_IMPORT anyharness/crates/anyharness-lib/src/sessions/store/events.rs 1 transitional session event sanitizer still lives under live session event sink
 SESSION_STORE_LIVE_IMPORT anyharness/crates/anyharness-lib/src/sessions/store/notifications.rs 1 transitional raw notification sanitizer still lives under live session event sink
 CORE_DOMAIN_PRODUCT_IMPORT anyharness/crates/anyharness-lib/src/sessions/mcp_bindings/injection.rs 2 transitional session MCP injection still imports product MCP modules
